// snl-gms-mock/packages/common-model/src/ts/channel-segment/types.ts
export interface ChannelSegmentDescriptor {
  readonly channel: VersionReference<'name', Channel> | Channel;
  readonly startTime: number;
  readonly endTime: number;
  readonly creationTime: number;
}